+++ app/package.json
@@ -1,7 +1,7 @@
 {
   "private": true,
   "name": "insomnia",
-  "version": "5.0.10",
+  "version": "5.0.12",
   "productName": "Insomnia",
   "longName": "Insomnia REST Client",
   "description": "Debug APIs like a human, not a robot",
@@ -9,18 +9,17 @@
   "author": "Insomnia <support@insomnia.rest>",
   "main": "main.min.js",
   "dependencies": {
-    "electron-context-menu": "0.8.0",
+    "electron-context-menu": "0.9.0",
     "electron-squirrel-startup": "1.0.0",
     "hkdf": "0.0.2",
     "httpsnippet": "1.16.5",
-    "insomnia-importers": "1.3.1",
-    "jsonpath": "0.2.9",
+    "insomnia-importers": "1.3.3",
+    "jsonpath": "0.2.11",
     "mkdirp": "0.5.1",
     "nedb": "1.8.0",
     "node-forge": "0.7.0",
     "node-libcurl": "git://github.com/getinsomnia/node-libcurl.git#3fc2afea435f3548eedbef9c68d3fee642dfcddb",
     "raven": "1.1.2",
-    "request": "2.71.0",
     "srp-js": "0.2.0",
     "tough-cookie": "2.3.1",
     "vkbeautify": "0.99.1"
