+++ app/ui/css/components/modal.less
@@ -118,6 +118,7 @@
     flex-direction: row;
     align-items: center;
     justify-content: space-between;
+    padding: 1px;
   }
 
   .modal__footer > *:last-child {
