+++ app/ui/css/components/sidebar.less
@@ -169,7 +169,7 @@
   .sidebar__list-root {
     // Add some space above so it's not so squished
     border-top: 1px solid @hl-sm;
-    padding-top: @padding-xs;
+    padding-top: @padding-xxs;
     padding-bottom: @padding-md;
     box-shadow: inset 0 2rem 2rem -2rem rgba(0, 0, 0, 0.03);
