+++ package.json
@@ -67,11 +67,13 @@
   "build": {
     "appId": "com.insomnia.app",
     "category": "public.app-category.developer-tools",
-    "protocols": [{
-      "name": "Insomnia",
-      "role": "Viewer",
-      "schemes": ["insomnia"]
-    }],
+    "protocols": [
+      {
+        "name": "Insomnia",
+        "role": "Viewer",
+        "schemes": ["insomnia"]
+      }
+    ],
     "fileAssociations": [],
     "directories": {
       "app": "build",
@@ -109,13 +111,13 @@
     "classnames": "2.2.5",
     "clone": "2.1.0",
     "codemirror": "5.24.2",
-    "electron-context-menu": "0.8.0",
+    "electron-context-menu": "0.9.0",
     "electron-squirrel-startup": "1.0.0",
     "hkdf": "0.0.2",
     "html-entities": "1.2.0",
     "httpsnippet": "1.16.5",
     "iconv-lite": "0.4.15",
-    "insomnia-importers": "1.3.1",
+    "insomnia-importers": "1.3.3",
     "jsonlint": "1.6.2",
     "jsonpath": "0.2.11",
     "mime-types": "2.1.14",
@@ -134,7 +136,6 @@
     "react-tabs": "0.8.2",
     "redux": "3.3.1",
     "redux-thunk": "2.0.1",
-    "request": "2.74.0",
     "reselect": "2.5.4",
     "srp-js": "0.2.0",
     "tough-cookie": "2.3.1",
